refacto(client): split chat in small part
